last update

+++ leerroutes/module03-keer-op-keer/robotarm-python-2021-main/example11.py
@@ -4,19 +4,27 @@ robotArm = RobotArm('exercise 11')
 
 # Jouw python instructies zet je vanaf hier:
 
-for i in range(0, 8):
-    robotArm.moveRight()
-for i in range(0, 9):
+counter = 9
+
+for move in range(counter):
     robotArm.grab()
     kleur = robotArm.scan()
-    if kleur == "white":
+    if kleur == '':
+        break
+    if kleur == 'white':
         robotArm.moveRight()
         robotArm.drop()
-        robotArm.moveLeft()
+        robotArm.moveRight()
     else:
         robotArm.drop()
-    robotArm.moveLeft()
+        robotArm.moveRight()
+
 robotArm.wait()
+
 
 # Na jouw code wachten tot het sluiten van de window:
 robotArm.wait() 
+
+#Verplaats alle witte blokken één plek naar rechts. 
+
+#Let op, de blokken zijn iedere keer anders als je het programma start!
